Conflicts:
	app/assets/stylesheets/active_admin/components/_tables.css.scss
	app/assets/stylesheets/active_admin/mixins/_utilities.scss

+++ lib/active_admin/views/index_as_table.rb
@@ -92,16 +92,16 @@ module ActiveAdmin
     #
     class IndexAsTable < ActiveAdmin::Component
 
-      def build(page_config, collection)
+      def build(page_presenter, collection)
         table_options = {
           :id => active_admin_config.plural_underscored_resource_name,
           :sortable => true,
           :class => "index_table",
-          :i18n => active_admin_config.resource
+          :i18n => active_admin_config.resource_class
         }
 
         table_for collection, table_options do |t|
-          instance_exec(t, &page_config.block)
+          instance_exec(t, &page_presenter.block)
         end
       end
 
@@ -126,9 +126,16 @@ module ActiveAdmin
             :name => ""
           }.merge(options)
           column options[:name] do |resource|
-            links = link_to I18n.t('active_admin.view'), resource_path(resource), :class => "member_link view_link"
-            links += link_to I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link"
-            links += link_to I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'), :class => "member_link delete_link"
+            links = ''.html_safe
+            if controller.action_methods.include?('show')
+              links += link_to I18n.t('active_admin.view'), resource_path(resource), :class => "member_link view_link"
+            end
+            if controller.action_methods.include?('edit')
+              links += link_to I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link"
+            end
+            if controller.action_methods.include?('destroy')
+              links += link_to I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'), :class => "member_link delete_link"
+            end
             links
           end
         end
